+++ fen_cli/src/codegen/client/swift/templates/Api.swift
@@ -1,17 +1,27 @@
 import Foundation
 
-struct ApiClient {
-  var fetcher: Fetcher
+struct APIClient {
+  var fetcher: any Fetcher
 }
 
-struct Fetcher {
+protocol Fetcher: Sendable {
+  func get<T>(from path: String, sessionToken: String?) async throws -> Response<T>
+  func post<T: Decodable, U: Encodable>(
+    to path: String,
+    with body: U,
+    returning type: T.Type,
+    sessionToken: String?
+  ) async throws -> Response<T>
+}
+
+struct LiveFetcher: Fetcher {
   var endpoint: String
 
   let jsonEncoder = JSONEncoder()
   let jsonDecoder = JSONDecoder()
 
   func get<T>(from path: String, sessionToken: String?) async throws -> Response<T>
-    where T: Decodable {
+  where T: Decodable {
     let url = URL(string: self.endpoint + path)!
     var request = URLRequest(url: url)
     request.httpMethod = "GET"
@@ -25,10 +35,10 @@ struct Fetcher {
     let tag = try self.jsonDecoder.decode(ResponseType.self, from: data)
     if tag.type == "success" {
       let response = try self.jsonDecoder.decode(SuccessResponse<T>.self, from: data)
-      return .success(SuccessResponse(data: response.data))
+      return .success(response.data)
     } else {
       let response = try self.jsonDecoder.decode(FailureResponse.self, from: data)
-      return .failure(FailureResponse(message: response.message, status: response.status))
+      return .failure(message: response.message, status: response.status)
     }
   }
 
@@ -54,10 +64,10 @@ struct Fetcher {
     let tag = try self.jsonDecoder.decode(ResponseType.self, from: data)
     if tag.type == "success" {
       let response = try self.jsonDecoder.decode(SuccessResponse<T>.self, from: data)
-      return .success(SuccessResponse(data: response.data))
+      return .success(response.data)
     } else {
       let response = try self.jsonDecoder.decode(FailureResponse.self, from: data)
-      return .failure(FailureResponse(message: response.message, status: response.status))
+      return .failure(message: response.message, status: response.status)
     }
   }
 }
@@ -69,8 +79,8 @@ struct ResponseType: Decodable {
 }
 
 enum Response<T: Decodable & Sendable>: Sendable {
-  case success(SuccessResponse<T>)
-  case failure(FailureResponse)
+  case success(T)
+  case failure(message: String, status: Int)
 }
 
 struct SuccessResponse<T: Decodable & Sendable>: Decodable, Sendable {